+++ src/icon-button.ts
@@ -1,0 +1,1 @@
+export * from "./components/icon-button/index.js";